feat(CommitTemplates): add templates
This commit adds all the templates to be utilized whenever you want to
make a semantic commit.

By issuing the command `$ gc` following with a semantic type it will
open a commit message using a templated pre-determined on
`commit_templates` folder.

Ex: `$ gc feature` or `$ gc chore`

A script was created in order to handle and determine which template is
supposed to be used.